Removed DB configuration + moved all libs logic in product service + removed DI
