--- libs/db/DB.interface.ts
@@ -1,13 +1,0 @@
-interface DBInterface {
-    connect: {
-        ( ): Promise <any>
-    },
-    query: {
-        ( query: string ): Promise <any>
-    },
-    disconnect: {
-        ( ): Promise <any>
-    }
-}
-
-export { DBInterface };
