--- libs/logger/logger.interface.ts
@@ -1,6 +1,0 @@
-export interface LoggerInterface {
-    logServiceRequest: ( message: string ) => void
-    logDBRequest: ( message: string ) => void
-    logError: ( message: string ) => void
-}
-
