--- libs/types/index.ts
@@ -1,7 +1,0 @@
-// @ts-ignore
-const TYPES = {
-    DB: Symbol.for("DB"),
-    LOGGER: Symbol.for("LOGGER")
-};
-
-export { TYPES };
