+++ product-service/handler.ts
@@ -1,4 +1,4 @@
-import * as handlers from './lambdas';
+import * as handlers from './src/lambdas';
 
 export const getProductById = handlers.getProductById;
 export const getAllProducts = handlers.getAllProducts;
