+++ product-service/products.json
@@ -1,0 +1,66 @@
+[
+  {
+    "count": 4,
+    "description": "Улучшение проекта существующего кода» пролили свет на процесс рефакторинга, описав принципы и лучшие приёмы его осуществления.",
+    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80aa",
+    "price": 2.4,
+    "title": "Рефакторинг от М.Фаулера",
+    "logo": "https://r5.readrate.com/img/pictures/basic/792/792600/7926007/w800h317-76be96b9.jpg"
+  },
+  {
+    "count": 6,
+    "description": "«Программист-прагматик. Путь от подмастерья к мастеру» расскажет всё, что нужно знать человеку, начинающему свой путь в сфере IT-проектов. Почти культовая книга. Вы узнаете, как бороться с недостатками ПО, как создать диамичную, эффективую и адаптируемую программу, как сформировать успешную команду программистов.",
+    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a0",
+    "price": 10,
+    "title": "Программист-прагматик. Путь от подмастерья к мастеру. Эндрю Хант, Дэвид Томас",
+    "logo": "https://r2.readrate.com/img/pictures/basic/792/792600/7926008/w800h317-89405d1d.jpg"
+  },
+  {
+    "count": 7,
+    "description": "«Совершенный код. Мастер-класс» – обновлённое издание проверенного временем бестселлера. Книга, которая заставляет размышлять и помогает создать идеальный код. Причём неважно, новичок вы или профи, в этом издании точно найдёте информацию для роста и работы над своим проектом. ",
+    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a2",
+    "price": 23,
+    "title": "Совершенный код. Мастер-класс. Стив Макконнелл",
+    "logo": "https://r5.readrate.com/img/pictures/basic/792/792600/7926009/w800h317-da60182f.jpg"
+  },
+  {
+    "count": 12,
+    "description": "Книга «На пике. Как поддерживать максимальную эффективность без выгорания» особенно необходима программистам, которые привыкли с головой погружаться в работу, не следя за временем и растратой таких ресурсов, как силы и здоровье. ",
+    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a1",
+    "price": 15,
+    "title": "На пике. Как поддерживать максимальную эффективность без выгорания. Брэд Сталберг, Стив Магнесс",
+    "logo": "https://r2.readrate.com/img/pictures/basic/792/792601/7926010/w800h317-04d81319.jpg"
+  },
+  {
+    "count": 7,
+    "description": "Эту книгу интересно читать и начинающему, и бывалому программисту. Авторы понятно и с юмором рассказывают о том, что программирование – это во многом коммуникация. Стиль программирования, именование, комментирование, работа с чужим кодом – зачастую соглашения складываются именно там, где строгая регламентация на уровне языка программи. ",
+    "id": "7567ec4b-b10c-48c5-9345-fc73c48a80a2",
+    "price": 23,
+    "title": "Программирование без дураков. Катрин Пассиг, Йоханнес Яндер",
+    "logo": "https://r2.readrate.com/img/pictures/basic/792/792601/7926014/w800h317-a1bf3137.jpg"
+  },
+  {
+    "count": 8,
+    "description": "Обновлённое юбилейное издание легендарной книги «Паттерны проектирования». Многие проблемы программирования имеют свойство повторяться и дублировать друг друга. Разработчики всего мира решают совершенно идентичные задачи и находят похожие решения. Если вы не хотите изобретать велосипед, используйте готовые шаблоны (паттерны) проектирования, работе с которыми посвящена эта книга.",
+    "id": "7567ec4b-b10c-48c5-9345-fc73348a80a1",
+    "price": 15,
+    "title": "Head First. Паттерны проектирования. Эрик Фримен, Элизабет Фримен",
+    "logo": "https://r5.readrate.com/img/pictures/basic/792/792601/7926015/w800h317-8f4d4f25.jpg"
+  },
+  {
+    "count": 2,
+    "description": "Можете ли вы без особого труда изменить код и тут же получить ответную реакцию на внесённые изменения? Насколько понятен этот код? Если вы сталкиваетесь с унаследованным кодом, вам пригодится книга Майкла К. Физерса, посвящённая работе с ним. \n\nВ книге «Эффективная работа с унаследованным кодом» вы найдёте информацию о механизмах внесения изменений в программное обеспечение, переносе унаследованного кода в средства тестирования, о написании тестов, препятствующих внесению новых ошибок в код, о применении методов, подходящих для любого языка или платформы, с примерами кода на Java, C++, C и C#. Кроме того, вы узнаете, как точно выявить в коде места, где требуется внести изменения, и как работать с унаследованным кодом, который не является объектно-ориентированным, и многое другое.",
+    "id": "7567ec4b-b10c-48c5-9445-fc73c48a80a2",
+    "price": 23,
+    "title": "Эффективная работа с унаследованным кодом. Майкл К. Физерс",
+    "logo": "https://r2.readrate.com/img/pictures/basic/792/792600/7926006/w800h317-535b11eb.jpg"
+  },
+  {
+    "count": 3,
+    "description": "Это издание – сборник полезных советов, многие из которых на собственном опыте можно сформулировать только после долгих лет работы в отрасли. Тут нет описания конкретных технологий, алгоритмов и языков программирования – ценность «Программиста-фанатика» не в этом. Книга расскажет, как победить отсутствие мотивации, выбрать правильные приоритеты, понять психологию программирования, наладить отношения с руководством и коллегами и т. д.",
+    "id": "7567ec4b-b10c-45c5-9345-fc73c48a80a1",
+    "price": 15,
+    "title": "Программист-фанатик. Чед Фаулер",
+    "logo": "https://r5.readrate.com/img/pictures/basic/792/792601/7926011/w800h317-16272c40.jpg"
+  }
+]
