+++ product-service/src/lambdas/index.ts
