+++ product-service/tsconfig.json
@@ -10,7 +10,8 @@
     "esModuleInterop": true,
     "sourceMap": true,
     "target": "es2017",
-    "outDir": "lib"
+    "outDir": "lib",
+    "resolveJsonModule": true
   },
   "include": ["./**/*.ts"],
   "exclude": [
